requirements.txt
streamlit
openai
langchain
langchain-community
tiktoken
pymupdf4llm
faiss-cpu